+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for NeoFS Node
 
 ### Added
 - Policer's setting to the SN's application configuration (#2600)
+- Support of verified domains for the storage nodes (#2280)
 
 ### Fixed
 - `neofs-cli netmap netinfo` documentation (#2555)